requirements.txt
fastapi
uvicorn
playwright
scrapy
httpx
google-generativeai
python-dotenv
pydantic
apify-client
